requirements.txt
torch ~= 1.7.0
numpy ~= 1.19.5
pandas ~= 1.1.4
scikit-learn ~= 0.23.2
pyyaml ~= 5.4.1
torchcde ~= 0.2.5